Start implementing Information layer
Use a registry to store agent information for each node, this way nodes
can be ignored if they are not relevant.
Waiting only 1 second for all the nodes to connect is a race condition,
the solution would be waiting longer which isn't ideal or monitoring
how many nodes are connected with a heartbeat. The heartbeat however
won't tell us when there are no longer anymore nodes left to connect.