Add current project diary file to version control
I will be working on this file for quite a while since my next meeting
will only be in a few months. So I am force adding it to git to ensure
it does not get deleted while I am working on it. Once it has been
finished, this file can be deleted and the pdf version of it can be
committed instead.