Check if all other nodes are connected and ready
I am committing this code so that a record of it exists. This code does
function very well at all, it is subject to race conditions, and I have
never seen every node be able to return all the other nodes using
DFService. Instead, I am going to be using a file which stores the
addresses of nodes ahead of time and just loading that into the
Knowledge layer. This is unfortunate given that this functionality did
work in Elixir.