TEMP: Copy .m2 folder into Docker container
https://jade.tilab.com/ is down meaning the JADE maven repository is
inaccessible. There are 5 different options to fix this.

1. Copy dependencies from the host system into the Docker container.
2. Migrate to using the fork at https://jade-project.gitlab.io/
3. Set up a local repository to serve the maven files for JADE.
4. Change build system to build from a local jar bundle.
5. Wait for https://jade.tilab.com/ to come back online.

I've decided to go with Option 1 for now as it is the quickest to set
up. The fork in Option 2 seems to rely on some artefacts from the main
JADE website, and I couldn't get it to work during a quick test.
Options 3 and 4 are ruled out for now as being too much effort to set
up. Hopefully Option 5 will be possible and the website won't be down
forever, but for now I'm going to go with Option 1 even if it
increases build times significantly.